Move the files to the correct folder
